Updated documentation.
Revised documentation in a number of files. Noticed a few issues along
the way, denoted via TODOs and FIXMEs.
