+++ src/server/plugin/test.cpp
@@ -1,5 +1,5 @@
 #include "../ouroboros_server.h"
-#include "../callback.h"
+#include "../callback.hpp"
 #include "test.h"
 #include <server/plugin.h>
 #include <iostream>
