+++ src/server/rest.h
@@ -38,15 +38,53 @@ namespace ouroboros
 	class rest_request
 	{
 	public:
+		/**	Constructor
+		 *
+		 *	@param [in] apConn Pointer to a connection object supplied by the
+		 *		Mongoose server.
+		 *	@param [in] aUri URI received from the server.
+		 * 
+		 */
 		rest_request(mg_connection *apConn, const std::string& aUri);
+		
+		/**	Destructor.
+		 * 
+		 */
 		~rest_request();
+		
+		/**	Returns the type of HTTP request for the connection of this request.
+		 *
+		 * 	@returns The HTTP request type for the connection of this request.
+		 */
 		http_request_type getHttpRequestType() const;
+		
+		/**	Returns the type of REST request type for the given URI.
+		 *
+		 * 	@returns The HTTP request type for the given URI.
+		 */
 		rest_request_type getRestRequestType() const;
 		
+		/**	Returns the field the REST request targets, if any.
+		 *
+		 *	@returns The field the REST request targets, or empty if none.
+		 */
 		std::string getFields() const;
+		
+		/**	Returns the group the REST request targets, if any.
+		 * 
+		 *	@returns the group the REST request targets, or empty if none.
+		 */
 		std::string getGroups() const;
+		
+		//TODO Implement functionality for REST callbacks
 		//std::string getCustom();
 		
+		/**	Returns a pointer to the connection object associated with the
+		 *		request.
+		 *
+		 *	@returns A pointer to the connection object associated with the
+		 *		request.
+		 */
 		mg_connection *getConnection() const;
 		
 	private:
@@ -56,7 +94,7 @@ namespace ouroboros
 		std::string mFields;
 		
 		mg_connection *mpConnection;
-		//TODO custom
+		//TODO callback functionality
 	};
 }
